+++ crates/dwallet-classgroups-types/Cargo.toml
@@ -4,6 +4,7 @@ edition = "2021"
 version.workspace = true
 
 [dependencies]
+dwallet-rng.workspace = true
 merlin.workspace = true
 class_groups.workspace = true
 group.workspace = true
@@ -12,8 +13,6 @@ crypto-bigint.workspace = true
 serde.workspace = true
 anyhow.workspace = true
 bcs.workspace = true
-fastcrypto = { workspace = true, features = ["copy_key"] }
-rand_chacha = "0.9"
 schemars.workspace = true
 serde_json.workspace = true
 serde_derive = "1.0.217"
@@ -22,7 +21,6 @@ dwallet-mpc-types.workspace = true
 twopc_mpc.workspace = true
 flate2 = "1.0.35"
 rand.workspace = true
-base64.workspace = true
 
 [lints]
 workspace = true
