+++ crates/dwallet-rng/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "dwallet-rng"
+edition = "2021"
+version.workspace = true
+
+[dependencies]
+merlin.workspace = true
+group.workspace = true
+proof.workspace = true
+fastcrypto.workspace = true
+base64.workspace = true
+ika-types.workspace = true
+rand_chacha.workspace = true
+serde.workspace = true
+
+[lints]
+workspace = true
