+++ crates/ika/Cargo.toml
@@ -13,6 +13,7 @@ workspace = true
 anyhow.workspace = true
 bin-version.workspace = true
 bcs.workspace = true
+dwallet-rng.workspace = true
 dwallet-classgroups-types.workspace = true
 clap.workspace = true
 datatest-stable.workspace = true
